refactor trace store chunk read io
